requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
Django==5.1.7
djangorestframework==3.15.2
dnspython==2.7.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
logger==1.4
mongoengine==0.29.0
oauthlib==3.2.2
openai==1.45.1
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.8.0
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
tweepy==4.15.0
typing_extensions==4.12.2
urllib3==2.3.0
